fix: 🐛 Fix safe directory error inside containers
When running conventional commits cli inside a Docker container as root user, i.e., a user different from the owner of the folder, git might raise an error saying that the current folder is not a safe directory. To avoid this git config --global --add safe.directory command is used.